rework permission system

+++ src/api/delete/crossRecord.ts
@@ -1,13 +1,23 @@
 import {sql} from "~/server-only/db";
 import {xName} from "~/util";
-import { getUserSession } from "~/server-only/session";
+import { belongsTo } from "~/server-only/session";
 import { CrossRecordMutateProps } from "../insert/crossRecord";
+
+export const whoCanDeleteCrossRecord = (tableName: string) => {
+  if (tableName === 'person') {
+    return []
+  } else {
+    return ['invited']
+  }
+}
 
 export const deleteCrossRecord = async (
   params: CrossRecordMutateProps
 ) => {
   "use server"
-  const userSession = await getUserSession()
+  if (! await belongsTo(whoCanDeleteCrossRecord(
+    params.first ? params.a : params.b
+  ))) return
   const tableName = xName(params.a, params.b, params.first)
   const result = await sql`
   DELETE FROM ${sql(tableName)}
